fix(iac/aws): split the #1636 guard, drop UK spellings and em-dashes
CI's golangci-lint (misspell, locale US) failed on `recognise` in the guard
test. Fixed, along with two `authorises` in the Terraform descriptions and the
five em-dashes the changeset introduced.

The guard test had grown to 521 lines against the project's 500-line rule, so
the discovery and trust/identity classification move to
cross_account_sts_discovery_test.go and the assertions stay put. Same package,
no behaviour change: 262 and 268 lines.

Mutation harness re-run and retargeted at the new file, since a mutation result
stops being evidence once the tree underneath it moves. 18/18 still bite.
